requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
Flask_JWT_Extended==4.6.0
pytest==8.2.2
python-dotenv==1.0.1
SQLAlchemy==2.0.31
Werkzeug==3.0.3
